CLI-837 fix update notification logic architecture
